+++ packages/data/src/index.ts
@@ -1,0 +1,8 @@
+export {default as Table} from './components/Table/Table';
+export {default as TableScroll} from './components/Table/TableScroll';
+export {default as TableHeader} from './components/Table/TableHeader';
+export {default as TableBody} from './components/Table/TableBody';
+export {default as TableFooter} from './components/Table/TableFooter';
+export {default as TableRow} from './components/Table/TableRow';
+export {default as TableCell} from './components/Table/TableCell';
+export {default as TableCellText} from './components/Table/TableCellText';