+++ packages/styles/src/theme/element/table.ts
@@ -1,0 +1,11 @@
+import {EmblemUIThemeTableStyles} from './element.interface';
+
+const table: EmblemUIThemeTableStyles = {
+  height: {
+    small: '32px',
+    default: '40px',
+    large: '48px'
+  }
+};
+
+export default table;