+++ packages/styles/src/theme/font/font.ts
@@ -1,0 +1,7 @@
+import {EmblemUIThemeFont} from './font.interface';
+
+const font: EmblemUIThemeFont = {
+  family: '"Montserrat", "Roboto", "sans-serif"'
+};
+
+export default font;