+++ packages/styles/src/theme/interface/palette.interface.ts
@@ -15,6 +15,8 @@ export interface EmblemUIThemePaletteBackground {
 }
 
 export interface EmblemUIThemePaletteBorder {
+  lightest: string;
+  light: string;
   color: string,
   hover: string;
 }