+++ CHANGELOG.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 ### Added
 - [styles] `ThemeProvider` component
+- [data] `Table` component
+- [data] `TableScroll` component
+- [data] `TableHeader` component
+- [data] `TableBody` component
+- [data] `TableFooter` component
+- [data] `TableRow` component
+- [data] `TableCell` component
+- [data] `TableCellText` component
 
 ## [0.1.1] - 2020-05-24
 ### Added